Fix VS x32 build warnings

+++ src/shell/shell_cmds.cpp
@@ -602,13 +602,13 @@ void DOS_Shell::CMD_RENAME(char * args){
 				if (dot2==NULL) {
 					star=strchr(arg2,'*');
 					if (strchr(arg2,'?')) {
-						for (int i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-arg2:strlen(arg2)); i++) {
+						for (size_t i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-arg2:strlen(arg2)); i++) {
 							if (*(arg2+i)=='?'&&i<strlen(name))
 								*(arg2+i)=name[i];
 						}
 					}
 					if (star) {
-						if (star-arg2<(unsigned int)strlen(name))
+						if ((unsigned int)(star-arg2)<(unsigned int)strlen(name))
 							strcpy(star, name+(star-arg2));
 						else
 							*star=0;
@@ -626,13 +626,13 @@ void DOS_Shell::CMD_RENAME(char * args){
 					*dot2='.';
 					star=strchr(tname2,'*');
 					if (strchr(tname2,'?')) {
-						for (int i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-tname2:strlen(tname2)); i++) {
+						for (size_t i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-tname2:strlen(tname2)); i++) {
 							if (*(tname2+i)=='?'&&i<strlen(tname1))
 								*(tname2+i)=tname1[i];
 						}
 					}
 					if (star) {
-						if (star-tname2<(unsigned int)strlen(tname1))
+						if ((unsigned int)(star-tname2)<(unsigned int)strlen(tname1))
 							strcpy(star, tname1+(star-tname2));
 						else
 							*star=0;
@@ -643,13 +643,13 @@ void DOS_Shell::CMD_RENAME(char * args){
 						strcpy(text2, dot2+1);
 						star=strchr(text2,'*');
 						if (strchr(text2,'?')) {
-							for (int i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-text2:strlen(text2)); i++) {
+							for (size_t i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-text2:strlen(text2)); i++) {
 								if (*(text2+i)=='?'&&i<strlen(text1))
 									*(text2+i)=text1[i];
 							}
 						}
 						if (star) {
-							if (star-text2<(unsigned int)strlen(text1))
+							if ((unsigned int)(star-text2)<(unsigned int)strlen(text1))
 								strcpy(star, text1+(star-text2));
 							else
 								*star=0;
@@ -657,7 +657,7 @@ void DOS_Shell::CMD_RENAME(char * args){
 					} else {
 						strcpy(text2, dot2+1);
 						if (strchr(text2,'?')||strchr(text2,'*')) {
-							for (int i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-text2:strlen(text2)); i++) {
+							for (size_t i=0; i<(uselfn?LFN_NAMELENGTH:DOS_NAMELENGTH) && i<(star?star-text2:strlen(text2)); i++) {
 								if (*(text2+i)=='*') {
 									*(text2+i)=0;
 									break;
